+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.55.0] - 2018-07-19
 ----------------------
 ##### Added
 - Support for 'no-ticks' appearance in range widget.